renamed package + formatted according to google style guide
