--- app/src/main/java/tu/bp21/passwortmanager/DB/PasswordDatabase.java
@@ -1,9 +1,0 @@
-package tu.bp21.passwortmanager.DB;
-
-import androidx.room.Database;
-import androidx.room.RoomDatabase;
-
-@Database(entities =  {User.class, Password.class}, version = 1)
-public abstract class PasswordDatabase extends RoomDatabase {
-    public abstract PasswortDao getDao();
-}
