+++ app/src/main/java/tu/bp21/passwortmanager/db/PasswordDatabase.java
@@ -1,0 +1,11 @@
+package tu.bp21.passwortmanager.db;
+
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
+
+@Database(
+    entities = {User.class, Password.class},
+    version = 1)
+public abstract class PasswordDatabase extends RoomDatabase {
+  public abstract PasswortDao getDao();
+}
