+++ app/src/main/java/tu/bp21/passwortmanager/db/User.java
@@ -1,0 +1,19 @@
+package tu.bp21.passwortmanager.db;
+
+import androidx.annotation.NonNull;
+import androidx.room.Entity;
+import androidx.room.PrimaryKey;
+
+@Entity
+public class User {
+
+  @PrimaryKey @NonNull public String username;
+  public String email;
+  public String password;
+
+  public User(@NonNull String username, String email, String password) {
+    this.username = username;
+    this.email = email;
+    this.password = password;
+  }
+}
